moved info to its own file. lifted up props to top of the file
